+++ Emu/source/Camera/TargetCamera.cpp
@@ -11,10 +11,10 @@ namespace Engine
 	{}
 
 	// Bug: Camera is stuttery for certain values of m_smoothingFactor and bounds.
-	void TargetCamera::Update(const float interpolation)
+	void TargetCamera::Update(const double interpolation)
 	{
-		float targetX = Lerp(ptrCameraTarget->GetPhysicsBody()->GetCenterPrevX(), ptrCameraTarget->GetPhysicsBody()->GetCenterXInMeters(), interpolation);
-		float targetY = Lerp(ptrCameraTarget->GetPhysicsBody()->GetCenterPrevY(), ptrCameraTarget->GetPhysicsBody()->GetCenterYInMeters(), interpolation);
+		float targetX = Lerp(ptrCameraTarget->GetPhysicsBody()->GetCenterPrevX(), ptrCameraTarget->GetPhysicsBody()->GetCenterXInMeters(), (float)interpolation);
+		float targetY = Lerp(ptrCameraTarget->GetPhysicsBody()->GetCenterPrevY(), ptrCameraTarget->GetPhysicsBody()->GetCenterYInMeters(), (float)interpolation);
 
 		// Desired camera position based on the target's position
 		float desiredCameraTopLeftX = targetX - (m_sizeInMeters.X / 2.0f);
